Added Check for Similarity
In extract_next_links(), before tokenizing the page, it checks to see if it is too similar to previous pages.

The global variable lsh represents our knowledge base, and every page is checked against it--to make sure it is not already in our knowledge--before being added to it.

This is done using MinHash from datasketch, a library found online